Boolean field test
